406: Update back button
