+++ Circles/Constants.swift
@@ -13,7 +13,6 @@ import SwiftUI
 
 let APP_PREFIX = "org.futo"
 
-
 let ROOM_TAG_CIRCLES_SPACE_ROOT = "org.futo.space.root"
 let ROOM_TAG_MY_CIRCLES = "org.futo.space.circles"
 let ROOM_TAG_MY_GROUPS = "org.futo.space.groups"
@@ -109,69 +108,3 @@ let CIRCLES_COLORS = [
 ]
 
 let DEFAULTS_KEY_ENABLE_GALLERIES = "enable_photo_galleries"
-
-enum SystemImages: String {
-    // MARK: - Multiply times
-    case checkmark = "checkmark"
-    case checkmarkCircle = "checkmark.circle"
-    case checkmarkShield = "checkmark.shield"
-    case chevronRight = "chevron.right"
-    case clock = "clock"
-    case desktopcomputer = "desktopcomputer"
-    case ellipsisCircle = "ellipsis.circle"
-    case exclamationmarkShield = "exclamationmark.shield"
-    case exclamationmarkTriangle = "exclamationmark.triangle"
-    case eye = "eye"
-    case eyeSlash = "eye.slash"
-    case gearshapeFill = "gearshape.fill"
-    case ipad = "ipad"
-    case iphone = "iphone"
-    case key = "key"
-    case keyFill = "key.fill"
-    case launchLogoPurple = "launch-logo-purple"
-    case link = "link"
-    case lockFill = "lock.fill"
-    case lockRectangle = "lock.rectangle"
-    case minusCircleFill = "minus.circle.fill"
-    case pencil = "pencil"
-    case pencilCircleFill = "pencil.circle.fill"
-    case personCropCircleBadgePlus = "person.crop.circle.badge.plus"
-    case personFill = "person.fill"
-    case personFillXmark = "person.fill.xmark"
-    case personLineDottedPersonFill = "person.line.dotted.person.fill"
-    case photoArtframe = "photo.artframe"
-    case playCircle = "play.circle"
-    case plusBubbleFill = "plus.bubble.fill"
-    case plusCircleFill = "plus.circle.fill"
-    case questionmarkCircle = "questionmark.circle"
-    case squareAndArrowUp = "square.and.arrow.up"
-    case trash = "trash"
-    case xmark = "xmark"
-    case xmarkBin = "xmark.bin"
-    case xmarkCircle = "xmark.circle"
-    case xmarkCircleFill = "xmark.circle.fill"
-    case xmarkShield = "xmark.shield"
-    
-    // MARK: - Once
-    case bubbleLeftAndBubbleRightFill = "bubble.left.and.bubble.right.fill"
-    case bubbleRight = "bubble.right"
-    case circle = "circle"
-    case circlesHexagonpath = "circles.hexagonpath"
-    case eyeFill = "eye.fill"
-    case eyeSlashFill = "eye.slash.fill"
-    case gearshape = "gearshape"
-    case heart = "heart"
-    case heartFill = "heart.fill"
-    case lockSlashFill = "lock.slash.fill"
-    case minusMagnifyingglass = "minus.magnifyingglass"
-    case paperclip = "paperclip"
-    case person2SquareStack = "person.2.square.stack"
-    case personCropSquare = "person.crop.square"
-    case photoBadgePlusFill = "photo.badge.plus.fill"
-    case photoFillOnRectangleFill = "photo.fill.on.rectangle.fill"
-    case plusMagnifyingglass = "plus.magnifyingglass"
-    case rectangleStackPersonCrop = "rectangle.stack.person.crop"
-    case speakerSlash = "speaker.slash"
-    case xCircleFill = "x.circle.fill"
-    case xmarkIcloudFill = "xmark.icloud.fill"
-}
