+++ Circles/Utilities/CustomFonts.swift
@@ -1,0 +1,17 @@
+//
+//  CustomFonts.swift
+//  Circles
+//
+//  Created by Dmytro Ryshchuk on 7/25/24.
+//
+
+import SwiftUI
+
+struct CustomFonts {
+    static let inter14 = Font.custom("Inter", size: 14)
+    static let nunito12 = Font.custom("Nunito", size: 12)
+    static let nunito14 = Font.custom("Nunito", size: 14)
+    static let nunito20 = Font.custom("Nunito", size: 20)
+    static let outfit11 = Font.custom("Outfit", size: 11)
+    static let outfit14 = Font.custom("Outfit", size: 14)
+}
