+++ Circles/Utilities/SystemImages.swift
@@ -1,0 +1,79 @@
+//
+//  SystemImages.swift
+//  Circles
+//
+//  Created by Dmytro Ryshchuk on 7/25/24.
+//
+
+import Foundation
+
+enum SystemImages: String {
+    // MARK: - Multiply times
+    case checkmark = "checkmark"
+    case checkmarkCircle = "checkmark.circle"
+    case checkmarkShield = "checkmark.shield"
+    case chevronRight = "chevron.right"
+    case clock = "clock"
+    case desktopcomputer = "desktopcomputer"
+    case ellipsisCircle = "ellipsis.circle"
+    case exclamationmarkShield = "exclamationmark.shield"
+    case exclamationmarkTriangle = "exclamationmark.triangle"
+    case eye = "eye"
+    case eyeSlash = "eye.slash"
+    case gearshapeFill = "gearshape.fill"
+    case iconFilledArrowBack = "IconFilledArrowBack"
+    case ipad = "ipad"
+    case iphone = "iphone"
+    case key = "key"
+    case keyFill = "key.fill"
+    case launchLogoPurple = "launch-logo-purple"
+    case link = "link"
+    case lockFill = "lock.fill"
+    case lockRectangle = "lock.rectangle"
+    case minusCircleFill = "minus.circle.fill"
+    case pencil = "pencil"
+    case pencilCircleFill = "pencil.circle.fill"
+    case personCropCircleBadgePlus = "person.crop.circle.badge.plus"
+    case personFill = "person.fill"
+    case personFillXmark = "person.fill.xmark"
+    case personLineDottedPersonFill = "person.line.dotted.person.fill"
+    case photoArtframe = "photo.artframe"
+    case playCircle = "play.circle"
+    case plusBubbleFill = "plus.bubble.fill"
+    case plusCircleFill = "plus.circle.fill"
+    case questionmarkCircle = "questionmark.circle"
+    case squareAndArrowUp = "square.and.arrow.up"
+    case trash = "trash"
+    case xmark = "xmark"
+    case xmarkBin = "xmark.bin"
+    case xmarkCircle = "xmark.circle"
+    case xmarkCircleFill = "xmark.circle.fill"
+    case xmarkShield = "xmark.shield"
+    
+    // MARK: - Once
+    case acceptedCircle = "acceptedCircle"
+    case bubbleLeftAndBubbleRightFill = "bubble.left.and.bubble.right.fill"
+    case bubbleRight = "bubble.right"
+    case circle = "circle"
+    case circlesHexagonpath = "circles.hexagonpath"
+    case eyeFill = "eye.fill"
+    case eyeSlashFill = "eye.slash.fill"
+    case forwardArrow = "forwardArrow"
+    case gearshape = "gearshape"
+    case heart = "heart"
+    case heartFill = "heart.fill"
+    case launchCircleBackground = "launchcirclebackground"
+    case lockSlashFill = "lock.slash.fill"
+    case minusMagnifyingglass = "minus.magnifyingglass"
+    case page = "Page"
+    case paperclip = "paperclip"
+    case person2SquareStack = "person.2.square.stack"
+    case personCropSquare = "person.crop.square"
+    case photoBadgePlusFill = "photo.badge.plus.fill"
+    case photoFillOnRectangleFill = "photo.fill.on.rectangle.fill"
+    case plusMagnifyingglass = "plus.magnifyingglass"
+    case rectangleStackPersonCrop = "rectangle.stack.person.crop"
+    case speakerSlash = "speaker.slash"
+    case xCircleFill = "x.circle.fill"
+    case xmarkIcloudFill = "xmark.icloud.fill"
+}
