+++ Circles/Views/Timeline/VideoContentView.swift
@@ -83,7 +83,7 @@ struct VideoContentView: View {
                     let markdown = MarkdownContent(caption)
                     Markdown(markdown)
                         .frame(maxWidth: .infinity, alignment: .leading)
-                        .font(Font.custom("Inter", size: 14))
+                        .font(CustomFonts.inter14)
                 }
                 
                 switch status {
